refactored BulkWriteBatch
this commit increases the distance between WriteBatch and BulkWriteBatch
