+++ google/cloud/firestore_v1/async_batch.py
@@ -18,22 +18,11 @@
 from google.api_core import gapic_v1  # type: ignore
 from google.api_core import retry as retries  # type: ignore
 
-from google.cloud.firestore_v1.base_batch import BaseBulkWriteBatch, BaseWriteBatch
+from google.cloud.firestore_v1.base_batch import BaseWriteBatch
 
 
-class _AsyncContextManagerMixin:
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, exc_type, exc_value, traceback):
-        if exc_type is None:
-            await self.commit()
-
-
-class AsyncWriteBatch(_AsyncContextManagerMixin, BaseWriteBatch):
-    """Accumulate write operations to be sent in a batch. Use this over
-    `AsyncBulkWriteBatch` for lower volumes or when the order of operations
-    within a given batch is important.
+class AsyncWriteBatch(BaseWriteBatch):
+    """Accumulate write operations to be sent in a batch.
 
     This has the same set of methods for write operations that
     :class:`~google.cloud.firestore_v1.async_document.AsyncDocumentReference` does,
@@ -76,52 +65,9 @@ class AsyncWriteBatch(_AsyncContextManagerMixin, BaseWriteBatch):
 
         return results
 
+    async def __aenter__(self):
+        return self
 
-class AsyncBulkWriteBatch(_AsyncContextManagerMixin, BaseBulkWriteBatch):
-    """Accumulate write operations to be sent in a batch. Use this over `AsyncWriteBatch`
-    for higher volumes (e.g., via `BulkWriter`) and when the order of operations
-    within a given batch is unimportant.
-
-    This has the same set of methods for write operations that
-    :class:`~google.cloud.firestore_v1.async_document.AsyncDocumentReference` does,
-    e.g. :meth:`~google.cloud.firestore_v1.async_document.AsyncDocumentReference.create`.
-
-    Args:
-        client (:class:`~google.cloud.firestore_v1.async_client.AsyncClient`):
-            The client that created this batch.
-    """
-
-    def __init__(self, client) -> None:
-        super(AsyncBulkWriteBatch, self).__init__(client=client)
-
-    async def commit(
-        self, retry: retries.Retry = gapic_v1.method.DEFAULT, timeout: float = None
-    ) -> list:
-        """Writes the changes accumulated in this batch.
-
-        Write operations are not guaranteed to be applied in order and must not
-        contain multiple writes to any given document. Preferred over `commit`
-        for performance reasons if these conditions are acceptable.
-
-        Args:
-            retry (google.api_core.retry.Retry): Designation of what errors, if any,
-                should be retried.  Defaults to a system-specified policy.
-            timeout (float): The timeout for this request.  Defaults to a
-                system-specified value.
-
-        Returns:
-            List[:class:`google.cloud.proto.firestore.v1.write.BatchWriteResult`, ...]:
-            The write results corresponding to the changes committed, returned
-            in the same order as the changes were applied to this batch. A
-            write result contains an ``update_time`` field.
-        """
-        request, kwargs = self._prep_commit(retry, timeout)
-
-        save_response = await self._client._firestore_api.batch_write(
-            request=request, metadata=self._client._rpc_metadata, **kwargs,
-        )
-
-        self._write_pbs = []
-        self.write_results = results = list(save_response.write_results)
-
-        return results
+    async def __aexit__(self, exc_type, exc_value, traceback):
+        if exc_type is None:
+            await self.commit()
