+++ google/cloud/firestore_v1/bulk_batch.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Google LLC All rights reserved.
+# Copyright 2021 Google LLC All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,17 +13,24 @@
 # limitations under the License.
 
 """Helpers for batch requests to the Google Cloud Firestore API."""
-
 from google.api_core import gapic_v1  # type: ignore
 from google.api_core import retry as retries  # type: ignore
 
-from google.cloud.firestore_v1.base_batch import BaseWriteBatch
+from google.cloud.firestore_v1 import _helpers
+from google.cloud.firestore_v1.base_batch import BaseBatch
 
 
-class WriteBatch(BaseWriteBatch):
-    """Accumulate write operations to be sent in a batch. Use this over
-    `BulkWriteBatch` for lower volumes or when the order of operations
-    within a given batch is important.
+class BulkWriteBatch(BaseBatch):
+    """Accumulate write operations to be sent in a batch. Use this over `WriteBatch`
+    for higher volumes (e.g., via `BulkWriter`) and when the order of operations
+    within a given batch is unimportant.
+
+    Because the order in which individual write operations land in the database
+    is not guaranteed, `batch_write` RPCs can never contain multiple operations
+    to the same document. If calling code detects a second write operation to a
+    known document reference, it should first cut off the previous batch and
+    send it, then create a new batch starting with the latest write operation. 
+    In practice, the [Async]BulkWriter classes handle this.
 
     This has the same set of methods for write operations that
     :class:`~google.cloud.firestore_v1.document.DocumentReference` does,
@@ -35,12 +42,16 @@ class WriteBatch(BaseWriteBatch):
     """
 
     def __init__(self, client) -> None:
-        super(WriteBatch, self).__init__(client=client)
+        super(BulkWriteBatch, self).__init__(client=client)
 
     def commit(
         self, retry: retries.Retry = gapic_v1.method.DEFAULT, timeout: float = None
-    ) -> list:
-        """Commit the changes accumulated in this batch.
+    ):
+        """Writes the changes accumulated in this batch.
+
+        Write operations are not guaranteed to be applied in order and must not
+        contain multiple writes to any given document. Preferred over `commit`
+        for performance reasons if these conditions are acceptable.
 
         Args:
             retry (google.api_core.retry.Retry): Designation of what errors, if any,
@@ -49,26 +60,28 @@ class WriteBatch(BaseWriteBatch):
                 system-specified value.
 
         Returns:
-            List[:class:`google.cloud.proto.firestore.v1.write.WriteResult`, ...]:
+            List[:class:`google.cloud.proto.firestore.v1.write.BatchWriteResult`, ...]:
             The write results corresponding to the changes committed, returned
             in the same order as the changes were applied to this batch. A
             write result contains an ``update_time`` field.
         """
         request, kwargs = self._prep_commit(retry, timeout)
 
-        commit_response = self._client._firestore_api.commit(
+        save_response = self._client._firestore_api.batch_write(
             request=request, metadata=self._client._rpc_metadata, **kwargs,
         )
 
         self._write_pbs = []
-        self.write_results = results = list(commit_response.write_results)
-        self.commit_time = commit_response.commit_time
+        self.write_results = results = list(save_response.write_results)
 
         return results
 
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type is None:
-            self.commit()
+    def _prep_commit(self, retry: retries.Retry, timeout: float):
+        """Shared setup for async/sync :meth:`write`."""
+        request = {
+            "database": self._client._database_string,
+            "writes": self._write_pbs,
+            "labels": None,
+        }
+        kwargs = _helpers.make_retry_timeout_kwargs(retry, timeout)
+        return request, kwargs
