+++ tests/unit/v1/test_async_batch.py
@@ -163,141 +163,6 @@ class TestAsyncWriteBatch(aiounittest.AsyncTestCase):
         firestore_api.commit.assert_not_called()
 
 
-class TestAsyncBulkWriteBatch(aiounittest.AsyncTestCase):
-    """Tests the AsyncBulkWriteBatch.commit method"""
-
-    @staticmethod
-    def _get_target_class():
-        from google.cloud.firestore_v1.async_batch import AsyncBulkWriteBatch
-
-        return AsyncBulkWriteBatch
-
-    def _make_one(self, *args, **kwargs):
-        klass = self._get_target_class()
-        return klass(*args, **kwargs)
-
-    def test_constructor(self):
-        batch = self._make_one(mock.sentinel.client)
-        self.assertIs(batch._client, mock.sentinel.client)
-        self.assertEqual(batch._write_pbs, [])
-        self.assertIsNone(batch.write_results)
-
-    @pytest.mark.asyncio
-    async def _write_helper(self, retry=None, timeout=None):
-        from google.cloud.firestore_v1 import _helpers
-        from google.cloud.firestore_v1.types import firestore
-        from google.cloud.firestore_v1.types import write
-
-        # Create a minimal fake GAPIC with a dummy result.
-        firestore_api = AsyncMock(spec=["batch_write"])
-        write_response = firestore.BatchWriteResponse(
-            write_results=[write.WriteResult(), write.WriteResult()],
-        )
-        firestore_api.batch_write.return_value = write_response
-        kwargs = _helpers.make_retry_timeout_kwargs(retry, timeout)
-
-        # Attach the fake GAPIC to a real client.
-        client = _make_client("grand")
-        client._firestore_api_internal = firestore_api
-
-        # Actually make a batch with some mutations and call write().
-        batch = self._make_one(client)
-        document1 = client.document("a", "b")
-        batch.create(document1, {"ten": 10, "buck": "ets"})
-        document2 = client.document("c", "d", "e", "f")
-        batch.delete(document2)
-        self.assertEqual(len(batch), 2)
-        write_pbs = batch._write_pbs[::]
-
-        write_results = await batch.commit(**kwargs)
-        self.assertEqual(write_results, list(write_response.write_results))
-        self.assertEqual(batch.write_results, write_results)
-        # Make sure batch has no more "changes".
-        self.assertEqual(batch._write_pbs, [])
-
-        # Verify the mocks.
-        firestore_api.batch_write.assert_called_once_with(
-            request={
-                "database": client._database_string,
-                "writes": write_pbs,
-                "labels": None,
-            },
-            metadata=client._rpc_metadata,
-            **kwargs,
-        )
-
-    @pytest.mark.asyncio
-    async def test_write(self):
-        await self._write_helper()
-
-    @pytest.mark.asyncio
-    async def test_write_w_retry_timeout(self):
-        from google.api_core.retry import Retry
-
-        retry = Retry(predicate=object())
-        timeout = 123.0
-
-        await self._write_helper(retry=retry, timeout=timeout)
-
-    @pytest.mark.asyncio
-    async def test_write_as_context_mgr_wo_error(self):
-        from google.cloud.firestore_v1.types import firestore
-        from google.cloud.firestore_v1.types import write
-
-        firestore_api = AsyncMock(spec=["batch_write"])
-        write_response = firestore.BatchWriteResponse(
-            write_results=[write.WriteResult(), write.WriteResult()],
-        )
-        firestore_api.batch_write.return_value = write_response
-        client = _make_client()
-        client._firestore_api_internal = firestore_api
-        batch = self._make_one(client)
-        document1 = client.document("a", "b")
-        document2 = client.document("c", "d", "e", "f")
-
-        async with batch as ctx_mgr:
-            self.assertIs(ctx_mgr, batch)
-            ctx_mgr.create(document1, {"ten": 10, "buck": "ets"})
-            ctx_mgr.delete(document2)
-            write_pbs = batch._write_pbs[::]
-
-        self.assertEqual(batch.write_results, list(write_response.write_results))
-        # Make sure batch has no more "changes".
-        self.assertEqual(batch._write_pbs, [])
-
-        # Verify the mocks.
-        firestore_api.batch_write.assert_called_once_with(
-            request={
-                "database": client._database_string,
-                "writes": write_pbs,
-                "labels": None,
-            },
-            metadata=client._rpc_metadata,
-        )
-
-    @pytest.mark.asyncio
-    async def test_write_as_context_mgr_w_error(self):
-        firestore_api = AsyncMock(spec=["batch_write"])
-        client = _make_client()
-        client._firestore_api_internal = firestore_api
-        batch = self._make_one(client)
-        document1 = client.document("a", "b")
-        document2 = client.document("c", "d", "e", "f")
-
-        with self.assertRaises(RuntimeError):
-            async with batch as ctx_mgr:
-                ctx_mgr.create(document1, {"ten": 10, "buck": "ets"})
-                ctx_mgr.delete(document2)
-                raise RuntimeError("testing")
-
-        # batch still has its changes, as _exit_ (and commit) is not invoked
-        # changes are preserved so commit can be retried
-        self.assertIsNone(batch.write_results)
-        self.assertEqual(len(batch._write_pbs), 2)
-
-        firestore_api.batch_write.assert_not_called()
-
-
 def _make_credentials():
     import google.auth.credentials
